refactor: extract GitHubExportProvider+Networking (#639 slice B)
Second slice of #639 (after #873 WireTypes at slice A). Sibling
of #874 (JiraExportProvider+Networking, PR #877) — same shape,
same pattern.

Moved 6 HTTP/URL/error helpers from GitHubExportProvider.swift
to a new Sources/BugNarrator/Services/GitHubExportProvider+Networking.swift
extension file:
- mapGitHubError(statusCode:data:configuration:retryAfterSeconds:) — private -> internal (widen)
- decodeGitHubMessage(from:) — stays private (single-caller, moves with)
- url(from components:) — private -> internal (widen)
- authenticatedRequest(url:httpMethod:token:includesJSONContentType:) — private -> internal (widen)
- repositoryEndpoint(configuration:) — private -> internal (widen)
- issueEndpoint(configuration:) — private -> internal (widen, called from makeURLRequest in base file)

Visibility widening: 5 of 6 helpers widen from private to internal
for cross-file reachability from remaining callers in the base
file. decodeGitHubMessage stays private because its sole caller
(mapGitHubError) moves into the same new extension file.

Codex spec-review verified all 5 to-widen helpers have cross-cluster
callers requiring internal; decodeGitHubMessage confirmed single-
caller. No external Sources/ or Tests/ file references any of the
6 items.

Line-count delta:
- GitHubExportProvider.swift: 734 -> 665 (-69 lines)
- GitHubExportProvider+Networking.swift (new): 77 lines
- xcodeproj: +4 (xcodegen)

Test coverage: GitHubExportProviderTests continues to pass unchanged.

Closes #878. Continues #639 after PR #875 (slice A WireTypes).
